First Version of Swipe Configuration in Saltstack
This is a working version of the Saltstack configuration for Swipe.
Currently only a few modules have been included, which are set to
install dependencies globally instead of in a virtual environment.
Several modules and pillar functionality will be added later.

This has been tested under salt minion, without running master. To
get the project working, you must install salt-minion and modify the
minion file in /etc/salt as follows:

     file_client: local

     file_roots:
       base:
         - /path/to/this/project/directory

You will need to uncomment these portions of the minion file and
specify the absolute path to the directory that this project was cloned
into to get this to work.