Python Virtualenv Added
This commit adds the beginning of a virtual environment to the
configuration. Currently, the the user is not specified (for the user
function and for the UNIX user directory). This will be later handled
through pillar when a pillar module is handled. As well, this only
installs virtualenv and creates an environment. It does not install
requirements.txt.