Fix for Virtual Environment
This commit cleans up the virtualenv installation, which was not working
when I attempted to install on a new machine. It seems that getting the
syntax for specifying user (runas) may have caused an issue, since it
worked after that was removed. The new syntax (user) has been added, but
not tested since it is currently commented out. As well this commit adds
some Jinja templating that is also not yet in use. The parts of the code
that are active appear to be working.