Node Intallation Working, Still Global
This commit fixes an issue that was keeping nodejs from installing
properly, and was preventing npm from installing the specified packages.
Since node is now packaged with npm, it appears that npm does not have
to be specified as a requisite. As of this commit, npm installs packages
globally. This needs to be fixed to install them into the project only.