Adds Bootstrap Script for Salt Minion Installation
This adds the salt bootstrap script that installs the saltstack. This is
the script that is recommended for installation, but since I have had
issues downloading and installing it on Ubuntu, I included a copy here.
Remember to run as superuser and to modify your salt minion file to
allow salt to run without salt master.