Installs Swipe Framework in Subdirectory
This installs the swipe-framework project from git in a directory called
'swipe' within the salt-framework directory. This setup isn't right,
though. The proper way to handle this is probably to install the git
project, and then add the virtual environment, requirements file, node
and ruby packages to the swipe-framework directory.