Properly Installs Project and Modfies Directory
This version creates a folder for the swipe-framework project and makes
changes to the project directory. Earlier, the git repo was unable to
install because the directory had already been modified, but this
appears to handle that problem by creating the folder first and
specifying it as a requirement. It may just be luck at this point that
it works at all, so keep an eye on what is being installed as the
project continues.