Git Requirement Properly Defined
This commit gets rid of the swipe-framework directory, which was being
used as a dependency for several states in an attempt to ensure that
things would be installed in the right order. I'm not entirely sure that
it was working properly in that form, but now the git project has been
added as a dependency instead. This should ensure that the project is
downloaded and then other files are installed into the project
directory.