Basic Nginx State - Install Script
This adds an nginx state, which only includes a python script that
installs and builds the special version of nginx that we need for the
project. As well, it generates a conf file and downloads a upload module
and debugging module from git. It might be better to move at least these
parts, if not the whole script to a salt state in the future, if
possible.