New Nginx Install Script
This replaces the python nginx installation script with a bash script.
The bash script seems to work as expected and is able to compile nginx
with the needed plugins. As well the conf file is now a managed file in the nginx state. This salt state should likely be changed to do some
of the work being done by the bash script for a cleaner installation. It
might be cleaner to have salt create the nginx directory, install the
conf file into it, and then run the script. As well, having salt handle
the cloned git projects might be worth doing.