Nginx Install Largely Moved to Salt
This commit takes much of the functionality out of the bash script that
installs nginx and moves it to the saltstate for nginx. This includes
all functionality that creates directories and clones git projects. It
also switches to the proper branch for the nginx upload module. As of
this commit, the bash script has been modified to remove these tasks,
but has not been fully tested.