Branch Checkout Added - Not Active
This adds a small change to the git salt state that specifies a branch
that is to be checked out for the swipe-framework project. It is still
commented out because something on that branch appears to be causing
problems. Because it is the current main branch for development, this
should be added in the future.