Ruby Gems Handled Through Gemfile
This changes the way that Ruby gems are installed in the ruby state. It
appears that RVM and the gems for the project were being installed
globally. This now installs a Gemfile that should handle the
installation, which was moved to salt from the swipe-framework git
project and is handled through the project's fabfile.