Adds NPM Back to Node State
This adds npm (Node Package Manager) back into the state for node. It
was removed previously because recent versions node include npm by
default. If an older version happens to get installed, then the state
will be missing npm, however. To avoid this it was added back into the
state.